+++ Annotations/SyncEntity.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace NTI\SyncBundle\Annotations;
+use Doctrine\Common\Annotations\Annotation\Required;
+
+/**
+ * Class SyncParent
+ * @package NTI\SyncBundle\Annotations
+ * @Annotation
+ */
+class SyncEntity {
+
+}
