--- Interfaces/SyncEntityInterface.php
@@ -1,6 +1,0 @@
-<?php
-
-namespace NTI\SyncBundle\Interfaces;
-
-interface SyncEntityInterface {
-}
