+++ Resources/config/services.yml
@@ -1,6 +1,6 @@
 services:
     nti.sync.doctrine.listener:
-        class: NTI\SyncBundle\EventListener\DoctrineEventSubscriber
+        class: NTI\SyncBundle\EventSubscriber\DoctrineEventSubscriber
         arguments: ["@service_container"]
         tags:
             - { name: doctrine.event_subscriber, connection: default }
